requirements.txt
langchain
langchain-chroma
langchain-groq
langchain-huggingface
langchain-text-splitters
python-dotenv
sentence-transformers
pypdf